+++ hw_4_logs_and_test/src/main/java/ua/com/alevel/service/impl/CustomerServiceImpl.java
@@ -3,6 +3,8 @@ package ua.com.alevel.service.impl;
 import ua.com.alevel.dao.CustomerDao;
 import ua.com.alevel.dao.impl.CustomerDaoImpl;
 import ua.com.alevel.entity.Customer;
+import ua.com.alevel.entity.CustomerAgreement;
+import ua.com.alevel.service.CustomerAgreementService;
 import ua.com.alevel.service.CustomerService;
 
 public class CustomerServiceImpl implements CustomerService {
@@ -21,7 +23,22 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void delete(String id) {
-        customerDao.delete(id);
+        CustomerAgreementService customerAgreementService = new CustomerAgreementServiceImpl();
+        CustomerAgreement[] customerAgreements = customerAgreementService.findAll();
+        boolean allowToDelete = true;
+        for (int i = 0; i < customerAgreements.length; i++) {
+            if (customerAgreements[i] != null && customerAgreements[i].getCustomer().getId().equals(id)) {
+                if(allowToDelete){
+                    System.out.println("For current company exist customer agreements, operation 'delete' isn't allowed!");
+                    System.out.println("You need to delete next agreements:");
+                    allowToDelete = false;
+                }
+                System.out.println(customerAgreements[i].getId());
+            }
+        }
+        if(allowToDelete) {
+            customerDao.delete(id);
+        }
     }
 
     @Override
